refactor: consolidate type definitions into single src/types.ts and remove electron/src/types.ts
