+++ electron/src/mcp/jsonrpc-handler.ts
@@ -2,7 +2,7 @@ import { McpServer } from '@modelcontextprotocol/sdk/server/mcp.js'
 import { z } from 'zod'
 import * as crypto from 'node:crypto'
 import { loadConfig } from '../config.js'
-import { McpLogItem, VERSION } from '../types.js'
+import { McpLogItem, VERSION } from '../../../src/types.js'
 import { confirmationManager, broadcastMcpEvent } from './confirmation-manager.js'
 import { recheckAuthorizationBeforeExecution, executeIsolatedSshCommand } from './ssh-executor.js'
 import { mcpExecutionManager } from './execution-manager.js'
